requirement.txt
streamlit==1.17.0
pandas==1.5.3
requests==2.28.2
beautifulsoup4==4.11.1
lxml==4.9.1
dnspython==2.2.1
smtplib==3.9.1
